FIxed an issue with explosions never being deleted
After a certain amount of time, explosions would gather up enough that
is caused slowdown; to reproduce, revert this change, spawn one tank,
then wait about a minute and try to move the view left or right.
Applying this change removes the issue; spawning a tank and lettign it
do its thang for a minute does not cause slowdown.